// src/partidas/dto/partida.dto.ts
export class PartidaDto {
    id: number;
    usuario1E1: string;
    usuario2E1: string;
    usuario3E2: string;
    usuario4E2: string;
    PontosU1E1: number;
    PontosU2E1: number;
    PontosU3E2: number;
    PontosU4E2: number;
    data: Date;
    

    placarEquipe1: number;
    placarEquipe2: number;

}